Files used to create vehicle
The chassis under design (2) is the design used to laser cut the bottom part of the vehicle. The Chassis Up Design is the design for the upper part of the vehicle. The Servo_axle is a 3d model to print the connection between the servo motor and a lego axle piece. Finally the pieces for chassis are the pieces needed to connect the ultrasonic sensor to the chassis.